In Json InspectorHandle use .toString() to get string value
Method .getTextValue() that was previously used returns null for
all that is not string. However, it makes sence to perform
conversion
to string from json map or array, when string type was requested.

For example:
SELECT val FROM tbl
LATERAL VIEW explode(json_split('[{"a": 4, "b": "f"}]')) x AS
json_val;

will output:
{"a": 4, "b": "f"}
as expected (but non "NULL" as before).